% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_item_in_any_doclib.R
\name{find_item_in_any_doclib}
\alias{find_item_in_any_doclib}
\title{find_item_in_any_doclib}
\usage{
find_item_in_any_doclib(site_url, target_filename, verbose = TRUE)
}
\arguments{
\item{site_url}{default is \code{NULL}. The base site URL.}

\item{target_filename}{default is \code{NULL}. The filename to look for.}

\item{verbose}{default is \code{TRUE}. Whether to display progress messages.}
}
\value{
List with found item and its drive/path.
}
\description{
Recursively searches all document libraries in a SharePoint site for a filename.
}
\details{
find_item_in_any_doclib
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
