% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_items_by_filename.R
\name{find_items_by_filename}
\alias{find_items_by_filename}
\title{find_items_by_filename}
\usage{
find_items_by_filename(
  drive,
  target_filename,
  parent_path = "",
  matches = list()
)
}
\arguments{
\item{drive}{default is \code{NULL}. Drive object to search.}

\item{target_filename}{default is \code{NULL}. Filename to look for.}

\item{parent_path}{default is \code{""}. Parent path to start in.}

\item{matches}{default is \code{list()}. Used internally during recursion. Ignore.}
}
\value{
List of found items, each with path and item.
}
\description{
Recursively searches a document library/drive for files matching a filename.
}
\details{
find_items_by_filename
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
