% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_m365_item.R
\name{find_m365_item}
\alias{find_m365_item}
\title{find_m365_item}
\usage{
find_m365_item(file_url, library = NULL)
}
\arguments{
\item{file_url}{default is \code{NULL}. The file's sharing URL.}

\item{library}{default is \code{NULL}. Optionally, the document library name.}
}
\value{
The found file item object, or error if not found.
}
\description{
Finds a file item on SharePoint/OneDrive from a submitted URL.
}
\details{
find_m365_item
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
