% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loader_for_extension.R
\name{get_loader_for_extension}
\alias{get_loader_for_extension}
\title{get_loader_for_extension}
\usage{
get_loader_for_extension(ext)
}
\arguments{
\item{ext}{default is \code{NULL}. The file extension (e.g., "csv", "xlsx").}
}
\value{
A function that loads the file into R if supported, otherwise downloads it.
}
\description{
get_loader_for_extension returns an appropriate handler function for a given file extension.

Supported file types and their behavior:
\itemize{
  \item \strong{csv, txt, tsv}: Loaded as data frames.
  \item \strong{xlsx, xls}: Loaded as data frames (first sheet by default).
  \item \strong{rds}: Loaded using \code{readRDS()}.
  \item \strong{rda, rdata}: Loaded into the global environment (\code{load()}).
}
All other file types (including \code{R}, \code{Rmd}, images, etc.) will be downloaded to the user's working directory.
}
\details{
get_loader_for_extension
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
