% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_m365_file.R
\name{load_m365_file}
\alias{load_m365_file}
\title{load_m365_file}
\usage{
load_m365_file(file_url, library = NULL, ...)
}
\arguments{
\item{file_url}{default is \code{NULL}. The file's full sharepoint/onedrive sharing URL.}

\item{library}{default is \code{NULL}. Optionally, the document library to search within (e.g. "Documents").}

\item{...}{Extra arguments passed to the relevant file loader (e.g., csv, excel, etc)}
}
\value{
The loaded R object, data.frame, or path to downloaded file.
}
\description{
Loads or downloads a SharePoint/OneDrive file given a sharing URL.
}
\details{
load_m365_file
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
