% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_doclib_name.R
\name{normalize_doclib_name}
\alias{normalize_doclib_name}
\title{normalize_doclib_name}
\usage{
normalize_doclib_name(site_url, doclib)
}
\arguments{
\item{site_url}{default is \code{NULL}. The base site URL.}

\item{doclib}{default is \code{NULL}. Library name.}
}
\value{
Character, possibly adjusted library name.
}
\description{
Normalizes the document library name for different SharePoint or OneDrive site types.
}
\details{
normalize_doclib_name
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
