% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_m365_url_components.R
\name{parse_m365_url_components}
\alias{parse_m365_url_components}
\title{parse_m365_url_components}
\usage{
parse_m365_url_components(file_url)
}
\arguments{
\item{file_url}{default is \code{NULL}. The sharing URL.}
}
\value{
List with \code{site_url}, \code{document_library_name}, \code{file_path_in_library}, and \code{library_unknown}.
}
\description{
Parses a SharePoint or OneDrive sharing URL into its site, library, and file path components.
}
\details{
parse_m365_url_components
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
